Fix test breakages: use `history.EffectType` where appropriate

+++ src/github.com/stellar/horizon/resource/effects/base.go
@@ -7,10 +7,12 @@ import (
 	"golang.org/x/net/context"
 )
 
+// PagingToken implements `db2.Pageable`
 func (this Base) PagingToken() string {
 	return this.PT
 }
 
+// Populate loads this resource from `row`
 func (this *Base) Populate(ctx context.Context, row history.Effect) {
 	this.ID = row.ID()
 	this.PT = row.PagingToken()
@@ -25,7 +27,7 @@ func (this *Base) Populate(ctx context.Context, row history.Effect) {
 
 func (this *Base) populateType(row history.Effect) {
 	var ok bool
-	this.TypeI = row.Type
+	this.TypeI = int32(row.Type)
 	this.Type, ok = TypeNames[row.Type]
 
 	if !ok {
